Create custom hook - Avatar.jsx의 fetch내용 custom hook으로 빼기
